fix: unblock GAME40001 research slot preview quality gate (Refs #3512)
Resolve the two ct_repo_lint failures on PR #3529:

- repo.app_event_bus_decoupling: the read-only ResearchFundingBreakdownDialog
  uses a local showDialog. Add research_slot_turn_preview_view.dart to the
  documented local-by-design allow-list (same rationale as the parent
  technology panel) and extend SPEC/program/app-ui-wiring.md accordingly.
- repo.game_widgets_file_size: technology_panel.dart grew to 726 lines. Extract
  the research slot-card widget family into technology_research_slot_card.dart
  (re-exported from technology_panel.dart so existing importers/tests keep
  resolving the moved symbols). Panel drops to 345 lines.
