+++ SPEC/program/app-ui-wiring.md
@@ -83,7 +83,7 @@ For `train_civilians` and `train_military`, shared order/count orchestration mus
 
 Combat-flow non-dialog screens (constructed directly by the orchestrator, not via `OpenDialogEvent`) are documented in [`quick-battle-screen.md`](../ui/quick-battle-screen.md), with sub-views [`quick-battle-deployment-view.md`](../ui/quick-battle-deployment-view.md) and [`quick-battle-action-selector.md`](../ui/quick-battle-action-selector.md).
 
-**Local by design (no `OpenDialogEvent`):** map display options (`GameMapArea`), tech detail (`TechTreeWidget`), civilian work-target sheet (`CivilianUnitsPanel`), next-turn confirmation (`GameMapArea`), in-game Android back exit confirmation (`GameScreen`), research tech picker (`TechnologyPanel`), **`CtDropdown`** internal picker, **new-game setup progress and error dialogs** after leader confirmation (see **SPEC/ui/game-initializing.md**), **`TransferToHomeFleetDialog`** (opened from `NavalUnitsPanel` via local `showDialog` for the regular-fleet → Home Fleet ship merge flow, commits via **`NavalTransferShipsRequestedEvent`**; see [SPEC/ui/transfer-to-home-fleet-dialog.md](../ui/transfer-to-home-fleet-dialog.md)), **`ProductionCommodityBreakdownDialog`** (opened from `ProductionScreen` via local `showDialog` as a read-only commodity preview; see [SPEC/ui/production-commodity-breakdown-dialog.md](../ui/production-commodity-breakdown-dialog.md)).
+**Local by design (no `OpenDialogEvent`):** map display options (`GameMapArea`), tech detail (`TechTreeWidget`), civilian work-target sheet (`CivilianUnitsPanel`), next-turn confirmation (`GameMapArea`), in-game Android back exit confirmation (`GameScreen`), research tech picker (`TechnologyPanel`), **`CtDropdown`** internal picker, **new-game setup progress and error dialogs** after leader confirmation (see **SPEC/ui/game-initializing.md**), **`TransferToHomeFleetDialog`** (opened from `NavalUnitsPanel` via local `showDialog` for the regular-fleet → Home Fleet ship merge flow, commits via **`NavalTransferShipsRequestedEvent`**; see [SPEC/ui/transfer-to-home-fleet-dialog.md](../ui/transfer-to-home-fleet-dialog.md)), **`ProductionCommodityBreakdownDialog`** (opened from `ProductionScreen` via local `showDialog` as a read-only commodity preview; see [SPEC/ui/production-commodity-breakdown-dialog.md](../ui/production-commodity-breakdown-dialog.md)), **`ResearchFundingBreakdownDialog`** (opened from the `TechnologyPanel` slot turn-preview view via local `showDialog` as a read-only RP/treasury funding breakdown, split out of `technology_panel.dart` to keep that file under the widget file-size cap; see [SPEC/ui/technology-panel.md](../ui/technology-panel.md) § Slot turn preview).
 
 **Split fleet:** `NavalUnitsPanel` uses local `showDialog` for `SplitFleetDialog`, but the dialog commits via **`NavalSplitFleetRequestedEvent`** → `AppEventHandlerScope` (applies `applyNavalSplitFleet`, then emits **`NavalFleetsUpdatedEvent`**) so the dialog does not receive panel merge callbacks. Widgetbook / tests without the shell wire the same request event or listen for `NavalFleetsUpdatedEvent` only.
